+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -448,6 +448,7 @@ sub build_isf_public_interface_contract {
             't/1279-isf-enum-member-inline-drive-values.t',
             't/1280-isf-enum-member-inline-drive-expression-values.t',
             't/1281-isf-enum-member-library-use-params.t',
+            't/1282-isf-enum-member-drive-expression-values.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',